requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
flask-cors==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5
gunicorn==21.2.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
Flask-WTF==1.2.1
reportlab==4.0.8
python-dateutil==2.8.2
